pyproject.toml
[project]
name = "feathers-cli"
version = "0.1.0"
description = "Scaffold production FastAPI services from YAML"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Abdul-Muizz Anwar", email = "abdulmuizz1310@outlook.com" }]
license = "MIT"
license-files = ["LICENSE"]
keywords = ["fastapi", "scaffold", "cli", "codegen"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "typer>=0.12.5",
    "rich>=13.9",
    "jinja2>=3.1.4",
    "pydantic>=2.9",
    "pyyaml>=6.0.2",
    "libcst>=1.5",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.13",
]

[project.urls]
Homepage = "https://github.com/Abdul-Muizz1310/feathers"
Repository = "https://github.com/Abdul-Muizz1310/feathers"
Issues = "https://github.com/Abdul-Muizz1310/feathers/issues"

[project.scripts]
feathers = "feathers.cli:app"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/feathers"]

[tool.hatch.build.targets.wheel.force-include]
"src/feathers/templates" = "feathers/templates"

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"src/feathers/cli.py" = ["B008"]  # Typer API requires function calls as defaults
"tests/**/*.py" = ["RUF043"]      # match= regex metacharacters are fine in tests

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
files = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: end-to-end tests that generate a service, sync deps, and boot it",
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20260408",
]
